Allow debugging the settings map access (#5)
This shows the settings map as a separate tab now. Also the statistics
are a separate tab now as well, because they show the handles in use as
well now. The auto splitters don't get reloaded on every settings change
anymore. It's currently unclear if we want to keep some sort of
backwards compatibility or not. For now you can simply press the
`Reload` button to force the auto splitter to be reloaded.