Update all the dependencies and clean up some code
This updates all the dependencies and cleans up some of the code. It
also improves the colors of the IDE. The main text color now has more
contrast and the log messages are now colored by their type. The layout
is also changed in preparation to change the entire debugger into a full
IDE in the future.